Convert Components to TypeScript
Sorry in advance, this was already a large edit upstream.  This had a cascade edit need due to reworking sortDirections and scrollDirection, which are heavily used.  The file count is high due to that.

I also had to fix a couple quirks with how MonitorToggleButton was refactored, to account for our customizations but also move it to TypeScript.  This also had a bunch of cascaded edit needs.

(cherry picked from commit sonarr/sonarr@e1cbc4a78249881de96160739a50c0a399ea4313)