Fix: Movie/Scene update performance (#851)
Recent changes to caching and resource mapping caused the PUT by ID API to slow down.  This was due to processing overhead in MapToResource, which was being called synchronously in tis API.  This presented as taking 3+ seconds to monitor/unmonitor a single movie.

MovieService.UpdateMovie is already triggering all the work asynchronously by triggering a MovieEditedEvent, so these calls inside MovieConroller were duplicates and expensive (MapToResource is heavy, so better in async).